backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
supabase==2.3.4
openai==1.12.0
requests==2.31.0
pydantic==2.6.0
python-multipart==0.0.9